task_2.1/requirements.txt
pytest==8.3.4
requests==2.32.3
allure-pytest==2.13.5
ruff==0.8.6
